python/requirements.txt
python-gitlab >= 3.15.0
